requirements.txt
kiteconnect
pandas
numpy
tvdatafeed
selenium
pyotp
python-dotenv